Refactor AkaRepository constructor for dependency injection
Modified the AkaRepository constructor to accept a NeutronDb object as a parameter, promoting dependency injection for better testability and flexibility. Removed the line initializing _context with a new NeutronDb instance and assigned the _context field the value of the context parameter. Commented out the Init method call within the constructor.