Refactor HistoryManager to use NeutronDb factory
Refactored the `HistoryManager` class in the `Neutron.Global` namespace to use a `Func<NeutronDb>` factory for creating instances of `NeutronDb` instead of direct instantiation. Updated the constructor to accept a `Func<NeutronDb>` parameter named `contextFactory`. Initialized private fields `_repoHistory`, `_repoReplenOrderDetails`, `_repoReplenOrders`, and `_repoInventory` using the `contextFactory` function. Added a null check for `contextFactory` in the constructor to ensure it is not null. Removed direct instantiation of `NeutronDb` in `GenericRepository` constructors, replacing it with the factory function approach.