Refactor FrmEditLocationDefinition to use context factory
Replaced the `_context` field with a `_contextFactory` of type `Func<NeutronDb>` in `FrmEditLocationDefinition.cs`. Updated the constructor to accept a `Func<NeutronDb>` parameter and throw an `ArgumentNullException` if it is null. Modified the instantiation of `GenericRepository` objects to use the factory method. These changes enhance flexibility and testability by allowing better dependency injection and easier unit testing.